Refactor device implementations into packages
