+++ src/heart/device/beats/__init__.py
@@ -1,0 +1,2 @@
+from heart.device.beats.device import StreamedScreen as StreamedScreen
+from heart.device.beats.websocket import WebSocket as WebSocket
