+++ src/heart/device/local/__init__.py
@@ -1,0 +1,3 @@
+from heart.device.local.device import LocalScreen as LocalScreen
+from heart.device.local.resolution import \
+    get_display_resolution as get_display_resolution
