+++ src/heart/device/single_led/__init__.py
@@ -1,0 +1,1 @@
+from heart.device.single_led.device import SingleLEDDevice as SingleLEDDevice
